fix: hooks for marker & embedding requests using correct ETag
