+++ src/api.v2/helpers/worker/workSubmit/submitEmbeddingWork.js
@@ -1,9 +1,9 @@
-const submitWork = require('./submitWork');
+const submitWorkForHook = require('./submitWorkForHook');
 
 const submitEmbeddingWork = async (message) => {
   const {
     experimentId, input:
-    { authJWT, config: { embeddingSettings: { method, methodSettings, useSaved } } },
+    { authJWT, config: { embeddingSettings: { method, methodSettings } } },
   } = message;
 
   const embeddingConfig = methodSettings[method];
@@ -11,17 +11,11 @@ const submitEmbeddingWork = async (message) => {
   const body = {
     name: 'GetEmbedding',
     type: method,
-    useSaved,
     config: embeddingConfig,
   };
 
-  // these values need to match explicitly the default ones defined in the UI at
-  // src/utils/work/fetchWork.js when calling the function generateETag if this file
-  // or the one in the UI has any default changed, the pre-computing of embeddings/marker heatmp
-  // will stop working as the ETags will no longer match.
-  const extraDependencies = [];
 
-  const ETag = await submitWork(experimentId, authJWT, body, extraDependencies);
+  const ETag = await submitWorkForHook(experimentId, authJWT, body);
 
   // explicitly return ETag to make it stand out more in tests and so harder to break
   return ETag;
