+++ src/api.v2/helpers/worker/workSubmit/submitMarkerHeatmapWork.js
@@ -1,8 +1,15 @@
-const getExtraDependencies = require('./getExtraDependencies');
-const submitWork = require('./submitWork');
+const getCellSetsAffectingDownsampling = require('./getCellSetsAffectingDownsampling');
+const submitWorkForHook = require('./submitWorkForHook');
+const bucketNames = require('../../../../config/bucketNames');
+const getObject = require('../../s3/getObject');
 
 const submitMarkerHeatmapWork = async (message) => {
   const { experimentId, input: { authJWT } } = message;
+
+  const { cellSets } = JSON.parse(await getObject({
+    Bucket: bucketNames.CELL_SETS,
+    Key: experimentId,
+  }));
 
   const body = {
     name: 'MarkerHeatmap',
@@ -15,9 +22,11 @@ const submitMarkerHeatmapWork = async (message) => {
     },
   };
 
-  const extraDependencies = await getExtraDependencies(experimentId, body.name, body);
+  const cs = await getCellSetsAffectingDownsampling(experimentId, body, cellSets);
 
-  const ETag = await submitWork(experimentId, authJWT, body, extraDependencies);
+  body.downsampleSettings.cellSets = cs;
+
+  const ETag = await submitWorkForHook(experimentId, authJWT, body);
 
   // explicitly return ETag to make it stand out more in tests and so harder to break
   return ETag;
