+++ tests/api.v2/helpers/worker/submitEmbeddingWork.test.js
@@ -1,13 +1,15 @@
+const AWSMock = require('aws-sdk-mock');
+
 const createObjectHash = require('../../../../src/api.v2/helpers/worker/createObjectHash');
 const submitEmbeddingWork = require('../../../../src/api.v2/helpers/worker/workSubmit/submitEmbeddingWork');
 const validateAndSubmitWork = require('../../../../src/api.v2/events/validateAndSubmitWork');
+const { mockS3GetObject } = require('../../../test-utils/mockAWSServices');
 
 
 jest.mock('../../../../src/api.v2/helpers/worker/createObjectHash');
 jest.mock('../../../../src/api.v2/helpers/pipeline/getPipelineStatus');
 jest.mock('../../../../src/api.v2/helpers/worker/getWorkerStatus');
 jest.mock('../../../../src/api.v2/events/validateAndSubmitWork');
-
 
 const message = {
   experimentId: '6463cb35-3e08-4e94-a181-6d155a5ca570',
@@ -44,7 +46,14 @@ const message = {
 };
 
 describe('submitWorkEmbedding', () => {
+  beforeEach(() => {
+    AWSMock.restore();
+    process.env.USE_CACHE = 'true';
+  });
+
   it('submits the work and the ETag / params are correct', async () => {
+    mockS3GetObject({ Body: '{}' });
+
     const ETag = await submitEmbeddingWork(message);
 
     expect(createObjectHash.mock.calls).toMatchSnapshot();
